CLN: Move stuff in tests.indexes.test_numeric.py to more logical locations (#41254)
